[REFACTOR/#220] 일기 관련 api들을 하나의 Service로 묶습니다.
