+++ app/src/main/java/com/sopt/clody/data/remote/datasourceimpl/DiaryListDataSourceImpl.kt
@@ -1,6 +1,5 @@
 package com.sopt.clody.data.remote.datasourceimpl
 
-import com.sopt.clody.data.remote.api.DiaryListService
 import com.sopt.clody.data.remote.datasource.DiaryListDataSource
 import com.sopt.clody.data.remote.dto.base.ApiResponse
 import com.sopt.clody.data.remote.dto.response.MonthlyDiaryResponseDto
