+++ app/src/main/java/com/sopt/clody/data/remote/datasourceimpl/MonthlyCalendarDataSourcelmpl.kt
@@ -1,6 +1,5 @@
 package com.sopt.clody.data.remote.datasourceimpl
 
-import com.sopt.clody.data.remote.api.CalendarApiService
 import com.sopt.clody.data.remote.datasource.MonthlyCalendarDataSource
 import com.sopt.clody.data.remote.dto.base.ApiResponse
 import com.sopt.clody.data.remote.dto.response.MonthlyCalendarResponseDto
