+++ app/src/main/java/com/sopt/clody/data/remote/datasourceimpl/ReplyDiaryDataSourceImpl.kt
@@ -1,6 +1,5 @@
 package com.sopt.clody.data.remote.datasourceimpl
 
-import com.sopt.clody.data.remote.api.ReplyDiaryService
 import com.sopt.clody.data.remote.datasource.ReplyDiaryDataSource
 import com.sopt.clody.data.remote.dto.response.ReplyDiaryResponseDto
 import com.sopt.clody.data.remote.dto.base.ApiResponse
